db: cast heat score sum to numeric for round(_, 2)
ln() returns double precision and Postgres has no
round(double precision, integer) overload, so applying
013_market_counter_triggers.sql aborted with:

  ERROR: function round(double precision, integer) does not exist

Cast the engagement-weight sum to numeric before round() so the
function definition compiles. No behaviour change beyond the
function being creatable.

+++ supabase/migrations/013_market_counter_triggers.sql
@@ -194,11 +194,15 @@ security definer
 set search_path = public
 as $$
   select round(
-    -- Engagement: log-scaled to dampen fat-tail bias
-    coalesce(ln(follow_count_denorm + 1) * 5.0, 0)
-      + coalesce(ln(target_count_denorm + 1) * 8.0, 0)
-      + coalesce(ln(grail_count_denorm  + 1) * 12.0, 0)
-      + coalesce(ln(owned_count_denorm  + 1) * 3.0, 0)
+    -- Engagement: log-scaled to dampen fat-tail bias.
+    -- Cast to numeric: ln() returns double precision and Postgres only
+    -- defines round(numeric, integer), not round(double precision, integer).
+    (
+      coalesce(ln(follow_count_denorm + 1) * 5.0, 0)
+        + coalesce(ln(target_count_denorm + 1) * 8.0, 0)
+        + coalesce(ln(grail_count_denorm  + 1) * 12.0, 0)
+        + coalesce(ln(owned_count_denorm  + 1) * 3.0, 0)
+    )::numeric
   , 2)
   from public.catalog_watch_market
   where catalog_watch_id = p_catalog_watch_id;
